Improves parallel test output, add --cores arg
